Refactored project: - Changed DeleteBehaviour after deleting Role - Changed Mapster config for entities with props with Stream type - Changed type of return value in CQRS for put/delete - Included `Storage.Net` for feature - Updated `ImageService` for normal working with directories - Changed type of assembly for including MediatR by DI - Fixed typo in not-nullable type extension - Added field for `ChangerId`. (only in test without authority) - Changed `ExecuteSqlRawAsync()` to `ExecuteUpdateAsync()` syntax

+++ MarketPlace.Infrastructure/Services/ImageService.cs
@@ -22,6 +22,11 @@ public class ImageService: IImageService
         {
             throw new ArgumentException("Invalid file extension");
         }
+        
+        if (!Directory.Exists(directory))
+        {
+            Directory.CreateDirectory(directory);
+        }
 
         var filePath = Path.Combine(directory, fileName);
         await using var fileStream = new FileStream(filePath, FileMode.Create);
@@ -49,10 +54,17 @@ public class ImageService: IImageService
             throw new ArgumentException("Invalid file extension");
         }
         
-        string tempFileName = "temp_" + fileName;
-        
-        await UploadImageAsync(image, directory, tempFileName);
-        await DeleteImageAsync(Path.Combine(directory, fileName));
-        File.Move(Path.Combine(directory,tempFileName), Path.Combine(directory, fileName));
+        if (!Directory.Exists(directory))
+        {
+            Directory.CreateDirectory(directory);
+            await UploadImageAsync(image, directory, fileName);
+        }
+        else
+        {
+            string tempFileName = "temp_" + fileName;
+            await UploadImageAsync(image, directory, tempFileName);
+            await DeleteImageAsync(Path.Combine(directory, fileName));
+            File.Move(Path.Combine(directory,tempFileName), Path.Combine(directory, fileName));
+        }
     }
 }
